move DEX contract utils/helpers to DexContracts;
